test: extract MockLogger
