+++ internal/project/application/usecases/shared_test.go
@@ -8,22 +8,6 @@ import (
 	"gomander/internal/eventbus"
 )
 
-type MockLogger struct {
-	mock.Mock
-}
-
-func (m *MockLogger) Info(message string) {
-	m.Called(message)
-}
-
-func (m *MockLogger) Debug(message string) {
-	m.Called(message)
-}
-
-func (m *MockLogger) Error(message string) {
-	m.Called(message)
-}
-
 type MockFsFacade struct {
 	mock.Mock
 }
